Fixes #2: Memory corruption in qtbrowserplugin fixed.
Google Chromium (and probably Chrome) frees the NPStream struct after
the NPP_DestroyStream method returns. But QtNPStream uses and accesses
references to NPStream afterwards. This leads to memory corruption and
finally to crashes of plugins.

The bug is not restricted to okularplugin. All other plugins using
qtbrowserplugin are affected by it.